+++ CHANGELOG.md
@@ -16,6 +16,7 @@ SPDX-License-Identifier: Apache-2.0
   which is no longer recommended, but still works.
   - Use a two-argument tag instead,
     which results in a partial application taking a table.
+- `hissp.apropos` to scan for imported symbols.
 
 ### Breaking